eal: remove unnecessary argv[0] handling
The rte_eal_init function looks at argv[0] to determine
the program name to pass to the log init function.
This is both unnecessary and in a corner case a problem.

Parsing argv[0] is unnecessary because the function openlog()
already determines the log identifier from program name if
NULL is passed, see openlog man page:
    The string pointed to by ident is prepended to every message, and is
    typically set to the program name. If ident is NULL, the program name
    is used.  (POSIX.1-2008 does not specify the behavior when ident is
    NULL.)

It is possible to run a program with argc=0 and argv[0]=NULL.
This would cause rte_eal_init() to dereference a null pointer.
Not a really useful feature, but better for DPDK get a SEGV
if it doesn't need to.

Signed-off-by: Stephen Hemminger <sthemmin@microsoft.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>