app/testpmd : fix testpmd quit error
when testpmd startup with pf and vfs, it is ok when running, while
exiting, it will result in heap-free-after-use which means pf is
released but vf is still accessing.

Change the logic of func port_is_bonding_slave ,this func
eth_dev_info_get_print_err while pf is released would result in this error.

Fixes: 0a0821bcf312 ("app/testpmd: remove most uses of internal ethdev array")
Cc: stable@dpdk.org

Signed-off-by: Wenxuan Wu <wenxuanx.wu@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>