eal: simplify argv[0] handling
The rte_eal_init function looks at argv[0] to determine
the program name to pass to the log init function.
But in corner cases argv[0] maybe NULL leading to a SEGV.

The code here is just using argv[0] to generate logid which
does not have to be in a static string, openlog() will handle
a const char pointer.

Simple workaround for argv[0] being NULL is to pass NULL
to openlog() and let it handle it. Both glibc, and musl
handle this case.

Signed-off-by: Stephen Hemminger <sthemmin@microsoft.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>