net/mlx5: add port flow configuration
The hardware steering is backend to support rte_flow_q API in mlx5
PMD. The port configuration function creates the queues and needed
flow management resources.

The PMD layer configuration function allocates the queues' context
and per-queue job descriptor. The job descriptor size is equal to
the queue size, and the job descriptors will be popped from LIFO
to convey the flow information during flow insertion/destruction.
So when polling the result, the flow information will be extracted
from the descriptor and then the job descriptor will be push back
to the LIFO.

The commit creates the flow port queue and the job descriptors.

Signed-off-by: Suanming Mou <suanmingm@nvidia.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>